Firmware/requirements.txt
python-can
ultralytics
wpilib
jupyterlab